Controller and View: merged their entity versions
Controller/Entity becomes Controller and View/Entity becomes View
These base classes reduce architectural complexity so there's less
classes to know/offer.
