+++ View.js
@@ -7,34 +7,67 @@ define([
   "use strict";
 
   var constructor = Class.extend({
-    init: function() {
-      this.on('postrender', function($view) {
-        if (typeof this.rendered === 'function') {
-          this.rendered($view);
+    init: function(model) {
+      this._model = model;
+
+      if (model && typeof model.id === 'function') {
+        this.id = function() {
+          return model.id && model.id();
+        };
+      }
+
+      this.on({
+        prerender: function() {
+          if (typeof this.prerender === 'function') {
+            this.prerender();
+          }
+        },
+        postrender: function($view) {
+          // Back-compat for view.rendered()
+          if (typeof this.rendered === 'function') {
+            this.rendered($view);
+          }
+          if (typeof this.postrender === 'function') {
+            this.postrender($view);
+          }
         }
       });
     },
-
-    render: function(data) {
-      var $existing = this.$view;
-
-      this.trigger('prerender', $existing);
-
-      this.$view = this.constructor.domify(this.template(data || this.templateData()));
-      this.constructor.replace($existing, this.$view);
-
-      this.trigger('postrender', this.$view);
-
-      return this.$view;
-    },
-
-    template: function() {},
-    templateData: function() { return {}; },
 
     destroy: function() {
       this.constructor.remove(this.$view);
       this.$view = null;
       this.off().stopListening();
+    },
+
+    template: function() {},
+
+    // All data needed to template the view
+    templateData: function() {
+      return (this._model && this._model.data) ? this._model.data() : this._model;
+    },
+
+    render: function($parent) {
+      var $existing = this.$view,
+          fresh = !($existing && $parent);
+
+      if (fresh) {
+        this.trigger('prerender', $existing);
+        this.$view = this.constructor.domify(this.template(this.templateData()));
+      }
+
+      if ($parent) {
+        this.constructor.appendTo(this.$view, $parent);
+      }
+      else {
+        this.constructor.replace($existing, this.$view);
+      }
+
+      if (fresh) {
+        this.trigger('postrender', this.$view);
+      }
+
+      return this.$view;
     }
   }, {
     displayName: 'View',
